requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-dotenv==0.19.0
twilio==7.11.0
openai==0.27.0
stripe==2.60.0
sqlalchemy==1.4.23
python-multipart==0.0.5
pydantic==1.8.2
requests==2.26.0
transformers==4.11.3
torch==1.9.0
